Updated Header, Drawer, and Footer
